Release of version V0.02
General fixes to the logic and trying to enforce that the site won't break easily.

Choosing a Penis Size or Breast Size will show options for type, setting the Size back to "N/A" will now hide the Type  options again.

Option to choose age has now been added. If input is 20s-90s, the prompt will now add in "Age: 20s" as an example. While a more specific age will add "27 years old" to the prompt. If age field is empty, prompt should be unaffected.